@@ -12,36 +12,45 @@
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@internationalized/date": "^3.7.0",
-    "@sveltejs/adapter-static": "^3.0.8",
-    "@sveltejs/kit": "^2.16.0",
-    "@sveltejs/vite-plugin-svelte": "^5.0.0",
+    "@internationalized/date": "^3.10.0",
+    "@sveltejs/adapter-static": "^3.0.10",
+    "@sveltejs/kit": "^2.48.4",
+    "@sveltejs/vite-plugin-svelte": "^5.1.1",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "^0.5.10",
-    "@tailwindcss/typography": "^0.5.16",
-    "autoprefixer": "^10.4.20",
-    "bits-ui": "^1.3.0",
+    "@tailwindcss/typography": "^0.5.19",
+    "autoprefixer": "^10.4.21",
+    "bits-ui": "^1.8.0",
     "clsx": "^2.1.1",
-    "embla-carousel-svelte": "^8.5.2",
-    "formsnap": "^2.0.0",
+    "embla-carousel-svelte": "^8.6.0",
+    "formsnap": "^2.0.1",
     "lucide-svelte": "^0.475.0",
     "mode-watcher": "^0.5.1",
-    "paneforge": "^1.0.0-next.4",
-    "svelte": "^5.0.0",
-    "svelte-check": "^4.0.0",
+    "paneforge": "^1.0.2",
+    "svelte": "^5.43.4",
+    "svelte-check": "^4.3.3",
     "svelte-sonner": "^0.3.28",
-    "sveltekit-superforms": "^2.23.1",
-    "tailwind-merge": "^3.0.1",
+    "sveltekit-superforms": "^2.28.1",
+    "tailwind-merge": "^3.3.1",
     "tailwind-variants": "^0.3.1",
-    "tailwindcss": "^3.4.17",
+    "tailwindcss": "^3.4.18",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "^5.0.0",
-    "vaul-svelte": "^1.0.0-next.6",
-    "vite": "^6.0.0",
-    "zod": "^3.24.2"
+    "typescript": "^5.9.3",
+    "vaul-svelte": "1.0.0-next.7",
+    "vite": "^6.4.1",
+    "zod": "^3.25.76"
   },
   "dependencies": {
+    "@supabase/ssr": "^0.7.0",
+    "@supabase/supabase-js": "^2.80.0",
+    "@sveltejs/adapter-auto": "^7.0.0",
     "@tanstack/table-core": "^8.21.3",
-    "embla-carousel-autoplay": "^8.5.2"
+    "@tiptap/core": "^3.10.2",
+    "@tiptap/extension-image": "^3.10.2",
+    "@tiptap/pm": "^3.10.2",
+    "@tiptap/starter-kit": "^3.10.2",
+    "diff": "^8.0.2",
+    "embla-carousel-autoplay": "^8.6.0",
+    "sveltekit-i18n": "^2.4.2"
   }
 }
