@@ -1,13 +1,31 @@
 // See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
+import { SupabaseClient, Session, User } from "@supabase/supabase-js";
+import { Database } from "$lib/database.types";
+
 declare global {
-	namespace App {
-		// interface Error {}
-		// interface Locals {}
-		// interface PageData {}
-		// interface PageState {}
-		// interface Platform {}
-	}
+  namespace App {
+    interface Error {
+      message: string;
+      code?: string;
+    }
+    interface Locals {
+      supabase: SupabaseClient<Database>;
+      safeGetSession: () => Promise<{
+        session: Session | null;
+        user: User | null;
+      }>;
+      session: Session | null;
+      user: User | null;
+      adminUser: { id: string; role: string } | null;
+    }
+    interface PageData {
+      session: Session | null;
+      user: User | null;
+    }
+    // interface PageState {}
+    // interface Platform {}
+  }
 }
 
 export {};
