@@ -1,1 +1,6 @@
-export const prerender = true
+// Removed prerender to support dynamic content from Supabase
+export const ssr = true;
+
+export const load = async ({ data }) => {
+  return data;
+};
